+++ src/app/employer/jobs/create/page.js
@@ -17,7 +17,7 @@ import BlueHyphonImg from '../../../../../public/assets/images/blue_hyphon.png';
 import PageTitleComponent from '@/components/employer/PageTitle/page_title';
 
 const JobCreatePage = () => {
-  const [activeStep, setActiveStep] = useState(1);
+  const [activeStep, setActiveStep] = useState(2);
   const [basicDetails, setBasicDetails] = useState({});
 
   useEffect(() => {
